Still working on age in 2000. Tinkering with what should count as a disturbance between 2000 and 2010 for setting the age in 2000.

+++ src/utilities/log_utilities.py
@@ -147,7 +147,10 @@ def merge_main_and_worker_upload_logs(no_log, main_log, worker_log, stage):
     s3_client = boto3.client("s3")  # Needs to be in the same function as the upload_file call
     s3_client.upload_file(combined_local_log, "gfw2-data", Key=f"{cn.s3_log_path}{combined_log_name}")
 
-    os.remove(main_log)
+    # Remove the main log if the stage doesn't run in batches.
+    # The main log must be kept for stages the run in batches because each batch uses the same main log.
+    if not "create_forest_age_2010_2015__1x1_deg" or "LULUCF_fluxes" in stage:
+        os.remove(main_log)
 
     if not no_log:
         os.remove(worker_log)
